feat(odsp-driver): support point-in-time loading in factory core (#28007)
Adds optional point-in-time loading support to
`OdspDocumentServiceFactoryCore`. The standard
`OdspDocumentServiceFactory` inherits this capability, while
`LocalOdspDocumentServiceFactory` leaves it undefined.

This removes the separate point-in-time factory implementation while
preserving the existing factory function, interface, and package
exports.

This change would make it easier for bohemia to consume point-in-time
load since `OdspDocumentServiceFactory` will support point-in-time
loading directly.

---------

Copilot-Session: 028eb841-af27-4aab-90b2-0b5a15609063
