+++ packages/drivers/odsp-driver/src/localOdspDriver/localOdspDocumentServiceFactory.ts
@@ -25,6 +25,8 @@ import { LocalOdspDocumentService } from "./localOdspDocumentService.js";
 export class LocalOdspDocumentServiceFactory extends OdspDocumentServiceFactoryCore {
 	private readonly logger: TelemetryLoggerExt = createOdspLogger();
 
+	public override readonly createPointInTimeDocumentService = undefined;
+
 	constructor(private readonly localSnapshot: Uint8Array | string) {
 		super(
 			(_options) => this.throwUnsupportedUsageError("Getting storage token"),
