+++ packages/drivers/odsp-driver/src/odspDocumentServiceFactoryCore.ts
@@ -19,6 +19,7 @@ import {
 } from "@fluidframework/driver-utils/internal";
 import {
 	type HostStoragePolicy,
+	type IOdspResolvedUrl,
 	type IOdspUrlParts,
 	type IRelaySessionAwareDriverFactory,
 	type ISharingLinkKind,
@@ -29,11 +30,20 @@ import {
 	type TokenFetchOptions,
 	type TokenFetcher,
 } from "@fluidframework/odsp-driver-definitions/internal";
-import { PerformanceEvent, createChildLogger } from "@fluidframework/telemetry-utils/internal";
+import {
+	PerformanceEvent,
+	UsageError,
+	createChildLogger,
+	type TelemetryLoggerExt,
+} from "@fluidframework/telemetry-utils/internal";
 import { v4 as uuid } from "uuid";
 
 import { useCreateNewModule } from "./createFile/index.js";
-import { type ICacheAndTracker, createOdspCacheAndTracker } from "./epochTracker.js";
+import {
+	type EpochTracker,
+	type ICacheAndTracker,
+	createOdspCacheAndTracker,
+} from "./epochTracker.js";
 import {
 	type INonPersistentCache,
 	type IPrefetchSnapshotContents,
@@ -41,6 +51,7 @@ import {
 	NonPersistentCache,
 } from "./odspCache.js";
 import { OdspDocumentService } from "./odspDocumentService.js";
+import { OdspDriverUrlResolver } from "./odspDriverUrlResolver.js";
 import { odspDriverCompatDetailsForLoader } from "./odspLayerCompatState.js";
 import {
 	type IExistingFileInfo,
@@ -52,6 +63,42 @@ import {
 	toInstrumentedOdspStorageTokenFetcher,
 	toInstrumentedOdspTokenFetcher,
 } from "./odspUtils.js";
+// eslint-disable-next-line import-x/no-internal-modules
+import { OdspPointInTimeDocumentService } from "./pointInTimeDriver/odspPointInTimeDocumentService.js";
+import {
+	createOdspVersionManager,
+	type IOdspVersionManager,
+} from "./odspVersionManager/index.js";
+
+/**
+ * An ODSP document service factory that supports point-in-time (sequence-number-based) loading.
+ *
+ * @remarks
+ * The loader detects this capability structurally, so hosts can pass this factory directly to
+ * {@link @fluidframework/container-loader#loadContainerToSequenceNumber}.
+ *
+ * @legacy @beta
+ */
+export interface IPointInTimeDocumentServiceFactory extends IDocumentServiceFactory {
+	/**
+	 * Creates a document service that materializes the document at the requested sequence number.
+	 *
+	 * @param resolvedUrl - The resolved ODSP {@link @fluidframework/driver-definitions#IResolvedUrl}.
+	 * @param targetSequenceNumber - The sequence number at which to materialize the document. See
+	 * {@link @fluidframework/container-loader#ILoadContainerToSequenceNumberProps.loadToSequenceNumber}.
+	 * @param logger - Optional {@link @fluidframework/core-interfaces#ITelemetryBaseLogger}.
+	 * @param clientIsSummarizer - Whether to apply summarizer policies and telemetry to the
+	 * underlying document services. Defaults to `false`.
+	 * @returns A read-only {@link @fluidframework/driver-definitions#IDocumentService} materialized
+	 * at the requested sequence number.
+	 */
+	createPointInTimeDocumentService(
+		resolvedUrl: IResolvedUrl,
+		targetSequenceNumber: number,
+		logger?: ITelemetryBaseLogger,
+		clientIsSummarizer?: boolean,
+	): Promise<IDocumentService>;
+}
 
 /**
  * Factory for creating the sharepoint document service. Use this if you want to
@@ -267,6 +314,134 @@ export class OdspDocumentServiceFactoryCore
 			undefined,
 			clientIsSummarizer,
 		);
+	}
+
+	/**
+	 * Creates a document service that reads its snapshot from the closest file version at or before
+	 * the target and its deltas from the live document, materializing a requested sequence number
+	 * through replay.
+	 *
+	 * @param resolvedUrl - The resolved ODSP {@link @fluidframework/driver-definitions#IResolvedUrl}.
+	 * @param targetSequenceNumber - The sequence number at which to materialize the document. See
+	 * {@link @fluidframework/container-loader#ILoadContainerToSequenceNumberProps.loadToSequenceNumber}.
+	 * @param logger - Optional {@link @fluidframework/core-interfaces#ITelemetryBaseLogger}.
+	 * @param clientIsSummarizer - Whether to apply summarizer policies and telemetry to the
+	 * underlying document services. Defaults to `false`.
+	 * @returns A read-only {@link @fluidframework/driver-definitions#IDocumentService} materialized
+	 * at the requested sequence number.
+	 */
+	public readonly createPointInTimeDocumentService?: IPointInTimeDocumentServiceFactory["createPointInTimeDocumentService"] =
+		async (
+			resolvedUrl: IResolvedUrl,
+			targetSequenceNumber: number,
+			logger?: ITelemetryBaseLogger,
+			clientIsSummarizer?: boolean,
+		): Promise<IDocumentService> => {
+			const odspLogger = createOdspLogger(logger);
+			const extLogger = createChildLogger({ logger: odspLogger });
+			const odspResolvedUrl = getOdspResolvedUrl(resolvedUrl);
+
+			// Use one epoch tracker for version selection, the base snapshot, and live ops so a
+			// point-in-time load cannot combine data from different file lineages.
+			const cacheAndTracker = createOdspCacheAndTracker(
+				this.persistedCache,
+				new NonPersistentCache(),
+				{
+					resolvedUrl: odspResolvedUrl,
+					docId: odspResolvedUrl.hashedDocumentId,
+					fileVersion: odspResolvedUrl.fileVersion,
+				},
+				extLogger,
+				clientIsSummarizer,
+			);
+
+			const versionManager = this.createVersionManager(
+				odspResolvedUrl,
+				extLogger,
+				cacheAndTracker.epochTracker,
+			);
+			const baseResult = await versionManager.findBaseForSeq(targetSequenceNumber);
+			if (baseResult.kind === "noBaseVersion") {
+				const oldestResolvedSequenceDetail =
+					baseResult.oldestResolvedSeq === undefined
+						? ""
+						: ` The oldest resolved file version is at sequence number ${baseResult.oldestResolvedSeq}.`;
+				throw new UsageError(
+					`No ODSP file version is available at or before sequence number ${targetSequenceNumber}.${oldestResolvedSequenceDetail}`,
+				);
+			}
+
+			const recoverableResolvedUrl = await this.resolveFileVersion(
+				resolvedUrl,
+				baseResult.base.versionId,
+			);
+			// Keep historical snapshots isolated from the normal factory cache while validating both
+			// services against the shared tracker.
+			const recoverableDocumentService = await this.createDocumentServiceCore(
+				recoverableResolvedUrl,
+				odspLogger,
+				cacheAndTracker,
+				clientIsSummarizer,
+			);
+			const liveDocumentService = await this.createDocumentServiceCore(
+				resolvedUrl,
+				odspLogger,
+				cacheAndTracker,
+				clientIsSummarizer,
+			);
+			return new OdspPointInTimeDocumentService(
+				recoverableResolvedUrl,
+				recoverableDocumentService,
+				liveDocumentService,
+				targetSequenceNumber,
+			);
+		};
+
+	/**
+	 * Creates the version manager used to select the closest file version at or before the target.
+	 */
+	private createVersionManager(
+		odspResolvedUrl: IOdspResolvedUrl,
+		logger: TelemetryLoggerExt,
+		epochTracker: EpochTracker,
+	): IOdspVersionManager {
+		const urlParts: IOdspUrlParts = {
+			siteUrl: odspResolvedUrl.siteUrl,
+			driveId: odspResolvedUrl.driveId,
+			itemId: odspResolvedUrl.itemId,
+		};
+		const getAuthHeader = toInstrumentedOdspStorageTokenFetcher(
+			logger,
+			urlParts,
+			this.getStorageToken,
+		);
+		return createOdspVersionManager({
+			urlParts,
+			getAuthHeader,
+			epochTracker,
+			logger,
+		});
+	}
+
+	private async resolveFileVersion(
+		resolvedUrl: IResolvedUrl,
+		fileVersion: string,
+	): Promise<IResolvedUrl> {
+		const odspResolvedUrl = getOdspResolvedUrl(resolvedUrl);
+		const query = new URLSearchParams({
+			driveId: odspResolvedUrl.driveId,
+			itemId: odspResolvedUrl.itemId,
+			fileVersion,
+		});
+		if (odspResolvedUrl.dataStorePath !== undefined) {
+			query.set("path", odspResolvedUrl.dataStorePath);
+		}
+		if (odspResolvedUrl.codeHint?.containerPackageName !== undefined) {
+			query.set("containerPackageName", odspResolvedUrl.codeHint.containerPackageName);
+		}
+		return new OdspDriverUrlResolver().resolve({
+			url: `${odspResolvedUrl.siteUrl}?${query.toString()}`,
+		});
 	}
 
 	protected async createDocumentServiceCore(
